requirements.txt
requests
beautifulsoup4
twilio
python-dotenv
pandas
numpy
selenium
webdriver-manager